move halfedge
