+++ src/compas/datastructures/halfedge/halfedge.py
