+++ src/compas/datastructures/mesh/clean.py
