+++ src/compas/datastructures/mesh/matrices.py
