+++ src/compas/datastructures/mesh/offset.py
@@ -5,8 +5,8 @@ from __future__ import print_function
 from compas.geometry import add_vectors
 from compas.geometry import scale_vector
 
-from compas.datastructures.mesh.orientation import mesh_flip_cycles
-from compas.datastructures.mesh.join import meshes_join
+from .orientation import mesh_flip_cycles
+from .join import meshes_join
 
 __all__ = [
     'mesh_offset',
