+++ src/compas/datastructures/mesh/operations/collapse.py
