+++ src/compas/datastructures/mesh/operations/merge.py
