+++ src/compas/datastructures/mesh/operations/split.py
