+++ src/compas/datastructures/mesh/operations/substitute.py
