+++ src/compas/datastructures/mesh/triangulation.py
@@ -3,7 +3,7 @@ from __future__ import absolute_import
 from __future__ import division
 
 
-from compas.datastructures.mesh.core import mesh_split_face
+from .operations import mesh_split_face
 
 
 __all__ = [
